Refactor artifacts-store
Change-type: patch
